Moving Sprinkle Manager
