+++ app/sprinkles/account/src/Account.php
@@ -11,7 +11,7 @@
 namespace UserFrosting\Sprinkle\Account;
 
 use UserFrosting\Sprinkle\Account\I18n\LocaleServicesProvider;
-use UserFrosting\System\Sprinkle\Sprinkle;
+use UserFrosting\Sprinkle\Core\Sprinkle\Sprinkle;
 
 /**
  * Bootstrapper class for the 'account' sprinkle.
