+++ app/sprinkles/core/tests/Unit/Sprinkle/data/sprinkles-bad.json
